scripts/requirements.txt
# For fetch_whitepaper.py
requests>=2.28.0
google-api-python-client>=2.100.0
google-auth>=2.22.0
google-auth-httplib2>=0.1.1
